Extract getPageFetchPromise method
