+++ src/modules/loadPage.ts
@@ -1,4 +1,4 @@
-import { classify, createHistoryRecord, fetch, getCurrentUrl } from '../helpers.js';
+import { classify, createHistoryRecord, getCurrentUrl } from '../helpers.js';
 import Swup from '../index.js';
 import { PageRecord } from './Cache.js';
 
@@ -42,32 +42,11 @@ const loadPage = function(this: Swup, data: TransitionOptions, popstate: PopStat
 		xhrPromise = this.preloadPromise;
 		this.preloadPromise = null;
 	} else {
-		// Fetch from server
-		xhrPromise = new Promise((resolve, reject) => {
-			fetch({ ...data, headers: this.options.requestHeaders }, (response) => {
-				if (response.status === 500) {
-					this.triggerEvent('serverError');
-					reject(url);
-					return;
-				}
-				// get json data
-				const page = this.getPageData(response);
-				if (!page || !page.blocks.length) {
-					reject(url);
-					return;
-				}
-				// render page
-				const cacheablePageData = { ...page, url };
-				this.cache.cacheUrl(cacheablePageData);
-				this.triggerEvent('pageLoaded');
-				resolve(cacheablePageData);
-			});
-		});
+		xhrPromise = this.getPageFetchPromise(data);
 	}
 
 	// when everything is ready, handle the outcome
-	// @ts-ignore two different promises should be fine to concat 🤷
-	Promise.all([xhrPromise].concat(animationPromises))
+	Promise.all([xhrPromise, ...animationPromises])
 		.then(([pageData]) => {
 			this.renderPage(pageData, { popstate, skipTransition });
 			this.preloadPromise = null;
